simplified the process of saving to the CalibrationOverlay.inc file
When INI settings were saved from the configuration server, differences were calculated
between what was being saved and what was in the default INI file and that was stored in
the CalibrationOverlay.inc. With the addition of feature overlays, as well, calculating
differences becomes a complicated problem. Instead of doing it that way, we now simply
show what is in the CalibrationOverlay.inc file and allow you to make additions/changes.
The full generated INI file that MachineKit is started with can be viewed in a new
INI Setup tab.

SOFT-175